% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyelinker.R
\docType{package}
\name{eyelinker-package}
\alias{eyelinker}
\alias{eyelinker-package}
\title{Import ASC Files from EyeLink Eye Trackers}
\description{
Dealing with unprocessed ASC files from EyeLink eye trackers can be a pain.
This package aims to make importing and working with these files as fast and
easy as possible.
}
\details{
For documentation of the structure of the returned data, see the "format" vignette:
\itemize{
\item \code{vignette("format", package = "eyelinker")}
}

For worked examples illustrating the package in action, see the "basics",
"epoching", and "batch_import" vignettes:
\itemize{
\item \code{vignette("basics", package = "eyelinker")}
\item \code{vignette("epoching", package = "eyelinker")}
\item \code{vignette("batch_import", package = "eyelinker")}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/a-hurst/eyelinker}
  \item Report bugs at \url{https://github.com/a-hurst/eyelinker/issues}
}

}
\author{
\strong{Maintainer}: Austin Hurst \email{austin.hurst@dal.ca}

Authors:
\itemize{
  \item Simon Barthelmé \email{simon.barthelme@gipsa-lab.fr}
}

}
