% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyelinker.R
\docType{package}
\name{eyelinker}
\alias{eyelinker}
\alias{eyelinker-package}
\title{eyelinker: read raw data from Eyelink eyetrackers}
\description{
Eyelink eye trackers output a horrible mess, typically under
 the form of an .asc file. The file in question is an assorted collection of
 messages, events and raw data. This R package will attempt to make sense of it.
}
\details{
The main function in the package is read.asc.
}

