% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%within\%}
\alias{\%within\%}
\alias{\%In\%}
\title{Find if value belongs to a set of intervals}
\usage{
x \%within\% Intv

x \%In\% Intv
}
\arguments{
\item{x}{A vector of numeric values}

\item{Intv}{A set of intervals, defined by a two-column matrix of endpoints or an Intervals
object}
}
\value{
A vector of logicals, which are true if x[i] belongs to any of the intervals in the set.
}
\description{
Returns whether numeric values on the left-hand side of the operator fall within any of the
specified intervals on the right-hand side. Intervals can be specified using either two-column
matrices or \code{\link[intervals:Intervals-class]{Intervals}} objects from the
\code{intervals} package.
}
\examples{
start <- c(0, 1, 2)
end <- c(.5, 1.3, 3)
intv <- cbind(start, end) # The 1st interval is 0.0-0.5, 2nd is 1.0-1.3, etc.
c(0, 0.6, 1.5, 3) \%within\% intv
}
\author{
Simon Barthelme
}
