% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyelink_parser.R
\name{read.asc}
\alias{read.asc}
\title{Read EyeLink ASC file}
\usage{
read.asc(fname, raw = TRUE, saccades = TRUE, fixations = TRUE,
  blinks = TRUE, msgs = TRUE, input = TRUE, buttons = TRUE)
}
\arguments{
\item{fname}{file name}
}
\value{
a list with components
  raw: raw eye positions, velocities, resolution, etc.
  msg: messages (no attempt is made to parse them)
  fix: fixations
  blinks: blinks
  sacc: saccades
  info: metadata
}
\description{
Read EyeLink ASC file
}
\details{
ASC files contain raw data from EyeLink eye trackers (they're ASCII versions of the raw binaries
which are themselves in EDF format). This utility tries to parse the data into something that's
usable in R. If all you have is an EDF file, you need to convert it first using the edf2asc
utility included in the EyeLink API kit.

Please read the EyeLink manual before using it for any serious work, very few checks are done to
see if the output makes sense. read.asc will return data frames containing a "raw" signal as
well as event series. Events are either system signals (triggers etc.), which are stored in the
"msg" field, or correspond to the EyeLink's interpretation of the eye movement traces
(fixations, saccades, blinks). ASC files are divided into blocks signaled by START and END
signals. The block structure is reflected in the "block" field of the dataframes. The names of
the various columns are the same as the ones used in the EyeLink manual, with two exceptions:
"cr.info", which gives you information about corneal reflection tracking, and "remote.info",
which gives you information about the state of the remote setup (if applicable). A series of
periods (e.g. ".....") for either indicates that tracking is functioning properly, other values
indicate problems with CR or remote tracking for that sample. Refer to the manual for details.
}
\examples{
# Example file from SR research that ships with the package
fpath <- system.file("extdata/mono500.asc.gz", package = "eyelinker")
dat <- read.asc(fpath)
plot(dat$raw$time, dat$raw$xp, xlab = "Time (ms)", ylab = "Eye position along x-axis (pix)")

}
\author{
Simon Barthelme
}
