% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyelinker.R
\docType{package}
\name{eyelinker}
\alias{eyelinker}
\alias{eyelinker-package}
\title{eyelinker: Import ASC Files from EyeLink Eye Trackers}
\description{
Dealing with unprocessed ASC files from EyeLink eye trackers can be a pain.
This package aims to make importing and working with these files as fast and
easy as possible.
}
\details{
The main function in the package is read.asc. Refer to the vignettes for
detailed information on the usage and data structure of this function.
}
