% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyelinker.R
\docType{package}
\name{eyelinker}
\alias{eyelinker}
\alias{eyelinker-package}
\title{Import ASC Files from EyeLink Eye Trackers}
\description{
Dealing with unprocessed ASC files from EyeLink eye trackers can be a pain.
This package aims to make importing and working with these files as fast and
easy as possible.
}
\details{
For documentation of the structure of the returned data, see the "format" vignette:

\code{vignette("format", package = "eyelinker")}

For worked examples illustrating the package in action, see the "basics" vignette:

\code{vignette("basics", package = "eyelinker")}
}
